+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/portlet/display/template/CPContentPortletDisplayTemplateHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.portlet.template;
+package com.liferay.commerce.product.content.web.internal.portlet.display.template;
 
 import com.liferay.commerce.product.constants.CPPortletKeys;
 import com.liferay.commerce.product.content.util.CPContentHelper;
